fix async use
