+++ ami/jobs/utils.py
@@ -1,0 +1,17 @@
+import asyncio
+import typing
+
+from ami.jobs.models import logger
+
+
+def _run_in_async_loop(func: typing.Callable, error_msg: str) -> typing.Any:
+    # helper to use new_event_loop() to ensure we're not mixing with Django's async context
+    loop = asyncio.new_event_loop()
+    asyncio.set_event_loop(loop)
+    try:
+        return loop.run_until_complete(func())
+    except Exception as e:
+        logger.error(f"Error in async loop - {error_msg}: {e}")
+        return None
+    finally:
+        loop.close()
